requirements.txt
langchain~=0.1.12
langchain-community~=0.0.38
pinecone-client~=5.0.1
langchain-openai~=0.1.23
openai
langchain_pinecone
uvicorn~=0.30.5
fastapi~=0.112.0
pydantic~=1.10.17
pypdf